Pushing my shit

+++ pintos/src/vm/frame.h
@@ -19,6 +19,6 @@ struct lock eviction_lock;
 void frame_init (size_t user_page_limit);
 size_t frame_alloc (enum palloc_flags flags);
 void frame_delete (size_t index);
-uint8_t *frame_get (size_t index);
+uint32_t *frame_get (size_t index);
 
 #endif /* vm/frame.h */
